Add files via upload
Install Node.js
install pug and run on localhost.
the web app will be always available as it is a free Azure technology which can be accessed here : https://flexweb.azurewebsites.net
The video Streaming is a paid Azure technology, hence that service will be stopped. The Perfect working model of the project can be seen through its ProjectDemo.